test-in-lxc: Fix source tree permissions after testing.
The provisioning phase runs as root and creates some files owned by root
in the source tree. For convenience, under lxc the tests themselves also
run as root using lxc-attach.

In order to not leave behind files owned by root, once the tests are
finished, we use lxc-attach in the container to reset all source
permissions to those of the test-in-lxc.sh script (which should be
unaltered). This will set them to the unprivileged user; once outside
the container, this should mean the files will be deletable by the user
who invoked test-in-lxc.sh.